+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 * [UICONSET-142](https://issues.folio.org/browse/UICONSET-142) Consortia-settings: format numbers in "Exported" and "Failed" columns.
 * [UICONSET-144](https://issues.folio.org/browse/UICONSET-144) Update characters limits of consortia member `code`.
 * [UICONSET-143](https://issues.folio.org/browse/UICONSET-143) Display toast message when deletion was failed for some members.
+* [UICONSET-44](https://issues.folio.org/browse/UICONSET-44) Implement consortium manager settings for "Call number types".